Move daemons to its own daemon package
